+++ saleor/payment/interface.py
@@ -79,6 +79,8 @@ class AddressData:
     country: str
     country_area: str
     phone: str
+    metadata: Optional[dict]
+    private_metadata: Optional[dict]
 
 
 class StorePaymentMethodEnum(str, Enum):